change long long to int64_t

+++ main.cc
@@ -23,6 +23,6 @@ int main(int argc, char **argv)
     if (d >= 64)
         return -1;
 
-    unsigned long long p = util::irred_poly(d);
+    uint64_t p = util::irred_poly(d);
     return 0;
 }
